mod topic part & add func
