+++ route/tool/func.py
@@ -882,7 +882,7 @@ def wiki_css(data):
     data += ['' for _ in range(0, 3 - len(data))]
     
     data_css = ''
-    data_css_ver = '146'
+    data_css_ver = '147'
     
     # Func JS
     data_css += '<script src="/views/main_css/js/func/func.js?ver=' + data_css_ver + '"></script>'
